refactor: Reservation 정보 List가 아닌 Object 형식으로 응답
