+++ src/main/java/roomescape/service/dto/response/reservation/ReservationResponse.java
@@ -1,7 +1,9 @@
-package roomescape.service.dto.response;
+package roomescape.service.dto.response.reservation;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
+import roomescape.service.dto.response.ReservationTimeResponse;
+import roomescape.service.dto.response.ThemeResponse;
 import roomescape.service.dto.response.member.MemberIdAndNameResponse;
 
 public record ReservationResponse(Long id,
